Добавлены репозитории для звонков и заявок

+++ src/bot/bot.update.ts
@@ -21,15 +21,14 @@ export class TgBotUpdate {
     @Message() msg: Record<string | symbol, any>,
   ) {
     try {
-       if(msg)
-        await ctx.deleteMessage(msg.id);
+      if (msg) await ctx.deleteMessage(msg.id);
       const testChatId = this.configService.get<string>('report.testChatId');
       const employersChatId = this.configService.get<string>(
         'report.employersChatId',
       );
       if (ctx.chat.id === +employersChatId) {
         await ctx.scene.enter('schedulleScene');
-        return
+        return;
       }
       // if (ctx.chat.id === +testChatId) {
       //   await ctx.scene.enter('schedulleScene');
